Avoid receiving incomplete UDP packages due to buffer overflow
When receiving UDP packet that is bigger than available buffer size
the excess bytes are discarded. Parser handling slows down due to errors
and it never reach to the point that there would be enough free space to
receive complete packets. The parser never recover from that unless recv
buffer is flushed to empty to make enough free space for complete packet.